Move and rename raycast method.
